refactor: move config to deno_json module
